pyproject.toml
[tool.poetry]
name = "musicplatform"
version = "0.1.0"
description = ""
authors = ["okhaled9 <okhaled9@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.2"
pytz = "^2023.3"
django-environ = "^0.10.0"
djangorestframework = "^3.14.0"
black = "^23.9.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
